feat: moved /app folder content to /[locale] and wrapped layout in NextIntlClientProvider(#15)
