+++ app/[locale]/(auth)/sign-up/loading.tsx
