+++ app/[locale]/(auth)/sign-up/page.tsx
