+++ app/[locale]/(protected)/[method]/[endpointBase64]/[bodyBase64]/loading.tsx
