+++ app/[locale]/(protected)/_components/body-editor.tsx
