+++ app/[locale]/(protected)/_components/codegener-panel.tsx
