+++ app/[locale]/(protected)/_components/headers-editor.tsx
