+++ app/[locale]/(protected)/_components/metrics-bar.tsx
