+++ app/[locale]/(protected)/_logic/encode.ts
