+++ app/[locale]/(protected)/_logic/normalize.ts
