+++ app/[locale]/(protected)/_logic/sync-url.ts
