+++ app/[locale]/(protected)/layout.tsx
