+++ app/[locale]/(public)/images/index.ts
@@ -1,0 +1,1 @@
+export { default as logoFull } from "./icons/Logo-paddle-text.svg";
