+++ app/[locale]/(public)/layout.tsx
