+++ app/[locale]/(public)/loading.tsx
