+++ app/[locale]/(public)/page.tsx
