+++ app/[locale]/_components/index.ts
@@ -1,2 +1,2 @@
-export { Header } from "./header";
+export { default as Header } from "./header";
 export { LanguageSwitch } from "./language-switch";
