+++ app/[locale]/api/history/route.ts
