+++ app/[locale]/layout.tsx
@@ -1,0 +1,31 @@
+import { ReactNode } from "react";
+import { NextIntlClientProvider } from "next-intl";
+import { getMessages, setRequestLocale } from "next-intl/server";
+
+import { Header } from "@app/[locale]/_components";
+
+import "@/shared/styles/globals.css";
+
+export default async function Layout({
+  children,
+  params,
+}: {
+  children: ReactNode;
+  params: Promise<{ locale: string }>; // <-- Next 15: params is async
+}) {
+  const { locale } = await params; // <-- await before use
+
+  setRequestLocale(locale);
+  const messages = await getMessages();
+
+  return (
+    <html lang={locale}>
+      <body className="bg-bg-primary text-base">
+        <NextIntlClientProvider locale={locale} messages={messages}>
+          <Header />
+          {children}
+        </NextIntlClientProvider>
+      </body>
+    </html>
+  );
+}
